draft rewrite updates screen in compose; draft dark theme; draft dynamic colors; draft in-app language; draft rewriting login screen; fix ic_launcher on older devices plus miui; etc

+++ app/src/main/kotlin/com/meloda/fast/screens/updates/model/UpdatesScreenState.kt
@@ -1,13 +1,12 @@
 package com.meloda.fast.screens.updates.model
 
 import com.meloda.fast.model.UpdateItem
-import com.meloda.fast.screens.updates.UpdateState
 
 data class UpdatesScreenState(
     val updateItem: UpdateItem?,
     val updateState: UpdateState,
     val error: String?,
-    val currentProgress: Int?,
+    val currentProgress: Float?,
     val isProgressIntermediate: Boolean,
 ) {
